update install.sh, fix shutdown fan 100% issue

+++ install.sh
@@ -5,6 +5,11 @@ cargo build --release;
 sudo rm /usr/local/bin/fan-manager;
 sudo cp target/release/fan-manager /usr/local/bin/fan-manager;
 sudo cp fan-manager.service /etc/systemd/system/fan-manager.service;
-sudo cp fan-manager.yaml /etc/fan-manager.yaml;
+
+# Only copy yaml if it doesn't exist
+if [ ! -f "/etc/fan-manager.yaml" ]; then
+    sudo cp fan-manager.yaml /etc/fan-manager.yaml;
+fi
+
 sudo systemctl daemon-reload;
 sudo systemctl enable --now fan-manager;
